designed views - EJS, added style
